Move filters to new module
