+++ Griffeye.VideoProcessor.Native.vcxproj
@@ -153,10 +153,12 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
+    <ClInclude Include="src\filters.h" />
     <ClInclude Include="src\filtergraph.h" />
     <ClInclude Include="src\libvx.h" />
   </ItemGroup>
   <ItemGroup>
+    <ClCompile Include="src\filters.c" />
     <ClCompile Include="src\filtergraph.c" />
     <ClCompile Include="src\libvx.c" />
   </ItemGroup>
